add project url

+++ database/migrations/2025_10_21_115602_modify_abouts_table_remove_video_button_url.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('abouts', function (Blueprint $table) {
-            $table->dropColumn('video_button_url');
+            if (Schema::hasColumn('abouts', 'video_button_url')) {
+                $table->dropColumn('video_button_url');
+            }
         });
     }
